moved athena_utils to filter_utils
